chore : service interface 추가
